Got the router parse to work.
Most of the router parsing elements work, and now stream into a structure
called a 'Loader'.  It is just a long array of content and types.  Can be
done with function pointers, or case or if or whatever floats your boat...

The only other modification is to save Lua functions as pointers (or userdata)
and recall them later for execution. The other option for functions would be to leave them
on the stack (and hope that it doesn't get GC'd), save the INDEX where the function
is on the stack, (as opposed to the address), and recall that with Lua's functions.